Broadcasting optionally includes the broadcaster
For replication phase, the leader needs to send himself "accept" so that he will also broadcast "accepted".
For both phases, the leader's default counts then start as 0
The option to send only to others is included so that upon election, a leader sends "I am leader" only to the other servers.